pyproject.toml
[project]
name = "reserve-model"
version = "0.1.0"
description = "HOA reserve planning spreadsheet generator"
requires-python = ">=3.11"
dependencies = [
  "openpyxl>=3.1",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["reserve"]
